// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefCusCodeListAttributeUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefCusCodeListAttributeUserView implements IEntity {
	ZZE_PK: string;
	ZZE_ZZD_CodeList: string;
	ZZE_ZXE_NKName: string;
	static readonly ZZE_ZXE_NKName_MaxLength: number = 32;
	ZZE_Value: string;
	static readonly ZZE_Value_MaxLength: number = 255;
	ZZE_StartDate: string | null;
	ZZE_EndDate: string | null;
	ZZE_IsAir: boolean;
	ZZE_IsSea: boolean;
	ZZE_IsFix: boolean;
	ZZE_IsRai: boolean;
	ZZE_IsRoa: boolean;
	ZZE_IsMai: boolean;
	ZZE_IsInw: boolean;
	ZZE_CodeType: string;
	static readonly ZZE_CodeType_MaxLength: number = 5;
	ZZE_CountryOrGrouping: string;
	static readonly ZZE_CountryOrGrouping_MaxLength: number = 3;
	ZZE_IsEditable: boolean;
}